[Complex] - 전투 상호작용 테스트
[Fix] - 상호작용 테스트를 위한 스크립트 디버그 및 간단한 필드값 수정

+++ Assets/01.Scripts/Util/Interfaces.cs
@@ -8,7 +8,7 @@ public interface IInteractable
 public interface IDamageable
 {
     void TakeDamage(float damage);
-    void BurnDamage(float damage, float attributeValue, float attributeRateTime, int attributeStack);
+    void BurnDamage(float damage, float attributeValue, float attributeRateTime, float attributeStack);
     void SlowDown(float damage, float attributeValue, float attributeRateTime);
 }
